refactor: remove duplication in item lists (inboxes)
