+++ packages/app/src/components/Inbox/index.js
@@ -1,0 +1,4 @@
+import ApproverInbox from './ApproverInbox';
+import RequestorInbox from './RequestorInbox';
+
+export { ApproverInbox, RequestorInbox };
